Consolidate multiple copies of StreamVacuum, make use UTF-8
